chore: add root npm scripts for lint and test
